+++ ml_tools/kerasmodel.py
@@ -13,6 +13,8 @@ from ml_tools.preprocess import (
 
 
 class KerasModel:
+    """ Defines a deep learning model using the tensorflow v2 keras framework """
+
     def __init__(self, train_config=None):
         self.params = {
             # augmentation
@@ -37,9 +39,7 @@ class KerasModel:
         if self.pretrained_model == "resnet":
             return (
                 tf.keras.applications.ResNet50(
-                    weights="imagenet",
-                    include_top=False,
-                    input_shape=input_shape,
+                    weights="imagenet", include_top=False, input_shape=input_shape,
                 ),
                 tf.keras.applications.resnet.preprocess_input,
             )
@@ -60,45 +60,35 @@ class KerasModel:
         elif self.pretrained_model == "vgg16":
             return (
                 tf.keras.applications.VGG16(
-                    weights="imagenet",
-                    include_top=False,
-                    input_shape=input_shape,
+                    weights="imagenet", include_top=False, input_shape=input_shape,
                 ),
                 tf.keras.applications.vgg16.preprocess_input,
             )
         elif self.pretrained_model == "vgg19":
             return (
                 tf.keras.applications.VGG19(
-                    weights="imagenet",
-                    include_top=False,
-                    input_shape=input_shape,
+                    weights="imagenet", include_top=False, input_shape=input_shape,
                 ),
                 tf.keras.applications.vgg19.preprocess_input,
             )
         elif self.pretrained_model == "mobilenet":
             return (
                 tf.keras.applications.MobileNetV2(
-                    weights="imagenet",
-                    include_top=False,
-                    input_shape=input_shape,
+                    weights="imagenet", include_top=False, input_shape=input_shape,
                 ),
                 tf.keras.applications.mobilenet_v2.preprocess_input,
             )
         elif self.pretrained_model == "densenet121":
             return (
                 tf.keras.applications.DenseNet121(
-                    weights="imagenet",
-                    include_top=False,
-                    input_shape=input_shape,
+                    weights="imagenet", include_top=False, input_shape=input_shape,
                 ),
                 tf.keras.applications.densenet.preprocess_input,
             )
         elif self.pretrained_model == "inceptionresnetv2":
             return (
                 tf.keras.applications.InceptionResNetV2(
-                    weights="imagenet",
-                    include_top=False,
-                    input_shape=input_shape,
+                    weights="imagenet", include_top=False, input_shape=input_shape,
                 ),
                 tf.keras.applications.inception_resnet_v2.preprocess_input,
             )
@@ -123,9 +113,7 @@ class KerasModel:
         preds = tf.keras.layers.Dense(len(self.labels), activation="softmax")(x)
         self.model = tf.keras.models.Model(inputs, outputs=preds)
         self.model.compile(
-            optimizer=self.optimizer(),
-            loss=self.loss(),
-            metrics=["accuracy"],
+            optimizer=self.optimizer(), loss=self.loss(), metrics=["accuracy"],
         )
 
     def loss(self):
@@ -216,10 +204,7 @@ class KerasModel:
         return output[0]
 
     def classify_track(
-        self,
-        clip,
-        track,
-        keep_all=True,
+        self, clip, track, keep_all=True,
     ):
         try:
             fp_index = self.labels.index("false-positive")
@@ -275,8 +260,10 @@ class KerasModel:
         frames_per_classify = self.square_width ** 2
         num_frames = len(data)
 
-        # note we can use more classifications but since we are using all regions
-        # with each classify for the over all movement, it doesn't change the result much
+        # note we can use more classifications but since we are using all track
+        # bounding regions with each classify for the over all movement, it
+        # doesn't change the result much
+        # take frames_per_classify random frames, sort by time then use this to classify
         num_classifies = math.ceil(float(num_frames) / frames_per_classify)
         frame_sample = np.arange(num_frames)
         np.random.shuffle(frame_sample)
@@ -288,17 +275,12 @@ class KerasModel:
             # update remaining
             frame_sample = frame_sample[frames_per_classify:]
             seg_frames.sort()
-            for i, frame_i in enumerate(seg_frames):
+            for frame_i in seg_frames:
                 f = data[frame_i]
                 segment.append(f)
 
             frames = preprocess_movement(
-                data,
-                segment,
-                self.square_width,
-                regions,
-                channel,
-                self.preprocess_fn,
+                data, segment, self.square_width, regions, channel, self.preprocess_fn,
             )
             if frames is None:
                 continue
